+++ CHANGES.md
@@ -21,9 +21,9 @@
 - [UPDATE] CUDA を 11.8 にあげる
   - 更新時に発生したビルド・エラーを回避するために `include/sora/fix_cuda_noinline_macro_error.h` を追加した
   - @enm10k
-- [UPDATE] SoraClientContextConfig, SoraVideoEncoderFactoryConfig に force_simulcast_i420_conversion を追加する
+- [UPDATE] SoraClientContextConfig, SoraVideoEncoderFactoryConfig に force_i420_conversion_for_simulcast_adapter を追加する
   - use_simulcast_adapter = true の際に、エンコーダー内でビデオ・フレームのバッファーを I420 に変換しているが、この変換の有無をフラグで制御可能にした
-  - force_simulcast_i420_conversion のデフォルト値は true で I420 への変換を行う
+  - force_i420_conversion_for_simulcast_adapter のデフォルト値は true で I420 への変換を行う
   - 変換を行わない場合、エンコードの性能が向上するが、バッファーの実装によってはサイマルキャストが利用できなくなる
   - @enm10k